task 10: route per-device callers through DeviceContextResolver and the moved DeviceSyncDispatcher
