+++ src/backend/src/agentclaw/community/core/channel/services/channel_service.py
@@ -23,11 +23,11 @@ from agentclaw.community.core.devices.services.device_context_resolver import (
 from agentclaw.community.di.modules.skill_center_module import (
     DeviceFilesystemDispatcher,
 )
-from agentclaw.community.core.devices.services.device_sync_dispatcher import DeviceSyncDispatcher
-from agentclaw.community.log import get_logger
-from agentclaw.community.plugin_api.device_sync import (
+from agentclaw.community.core.devices.services.device_sync import (
     DeviceSyncUnavailableError,
 )
+from agentclaw.community.plugin_api.device_sync_dispatcher import DeviceSyncDispatcher
+from agentclaw.community.log import get_logger
 
 logger = get_logger()
